+++ system/src/sheets/FalloutActorSheet.mjs
@@ -124,15 +124,12 @@ export default class FalloutActorSheet extends ActorSheet {
    * @return {undefined}
    */
 	_prepareCharacterData(context) {
-		// // Handle ability scores.
-		// for (let [k, v] of Object.entries(context.system.attributes)) {
-		// 	v.label = game.i18n.localize(CONFIG.FALLOUT.attributes[k]) ?? k;
-		// }
-
 		let allInjuries = [];
+
 		for (const [, bp] of Object.entries(this.actor.system.body_parts)) {
 			allInjuries.push(...bp.injuries);
 		}
+
 		context.treatedInjuriesCount = allInjuries.filter(i => i === 1).length;
 		context.openInjuriesCount = allInjuries.filter(i => i === 2).length;
 
@@ -154,91 +151,78 @@ export default class FalloutActorSheet extends ActorSheet {
    * @return {undefined}
    */
 	async _prepareItems(context) {
-		const skills = [];
-		const perks = [];
-		const apparel = [];
-		const apparel_mods = [];
-		const robotApparel = [];
-		const robot_mods = [];
-		const weapons = [];
-		const weapon_mods = [];
-		const ammo = [];
-		const consumables = [];
-		const books_and_magz = [];
-		const miscellany = [];
-		const specialAbilities = [];
-		const diseases = [];
+		context.itemsByType = {
+			ammo: [],
+			apparel_mod: [],
+			apparel: [],
+			books_and_magz: [],
+			consumable: [],
+			disease: [],
+			miscellany: [],
+			perk: [],
+			robot_armor: [],
+			robot_mod: [],
+			skill: [],
+			special_ability: [],
+			trait: [],
+			weapon_mod: [],
+			weapon: [],
+		};
 
+		const apparel = [];
+		const robotApparel = [];
 
 		// Iterate through items, allocating to containers
 		for (let i of context.items) {
 			i.img = i.img || DEFAULT_TOKEN;
 
-			if (i.type === "skill") {
-				i.localizedName = game.i18n.localize(
-					`FALLOUT.SKILL.${i.name}`
-				);
+			switch (i.type) {
+				case "ammo":
+					i.shotsAvailable = ((i.system.quantity - 1)
+					* i.system.shots.max
+					) + i.system.shots.current;
 
-				i.localizedDefaultAttribute = game.i18n.localize(
-					`FALLOUT.AbilityAbbr.${i.system.defaultAttribute}`
-				);
-
-				skills.push(i);
-			}
-			else if (i.type === "perk") {
-				perks.push(i);
-			}
-			else if (i.type === "apparel") {
-				apparel.push(i);
-			}
-			else if (i.type === "apparel_mod") {
-				apparel_mods.push(i);
-			}
-			else if (i.type === "robot_armor") {
-				robotApparel.push(i);
-			}
-			else if (i.type === "robot_mod") {
-				robot_mods.push(i);
-			}
-			else if (i.type === "weapon") {
-				if (i.system.ammo !== "") {
-					const [, shotsAvailable] = await this.actor._getAvailableAmmoType(
-						i.system.ammo
+					context.itemsByType.ammo.push(i);
+					break;
+				case "apparel":
+					apparel.push(i);
+					break;
+				case "robot_armor":
+					robotApparel.push(i);
+					break;
+				case "skill":
+					i.localizedName = game.i18n.localize(
+						`FALLOUT.SKILL.${i.name}`
 					);
 
-					i.shotsAvailable = shotsAvailable;
-				}
-				weapons.push(i);
-			}
-			else if (i.type === "weapon_mod") {
-				weapon_mods.push(i);
-			}
-			else if (i.type === "ammo") {
-				i.shotsAvailable =
-					((i.system.quantity - 1) * i.system.shots.max) + i.system.shots.current;
+					i.localizedDefaultAttribute = game.i18n.localize(
+						`FALLOUT.AbilityAbbr.${i.system.defaultAttribute}`
+					);
 
-				ammo.push(i);
-			}
-			else if (i.type === "consumable") {
-				consumables.push(i);
-			}
-			else if (i.type === "books_and_magz") {
-				books_and_magz.push(i);
-			}
-			else if (i.type === "miscellany") {
-				miscellany.push(i);
-			}
-			else if (i.type === "special_ability") {
-				specialAbilities.push(i);
-			}
-			else if (i.type === "disease") {
-				diseases.push(i);
+					context.itemsByType.skill.push(i);
+					break;
+				case "weapon":
+					if (i.system.ammo !== "") {
+						const [, shotsAvailable] =
+							await this.actor._getAvailableAmmoType(
+								i.system.ammo
+							);
+
+						i.shotsAvailable = shotsAvailable;
+					}
+					context.itemsByType.weapon.push(i);
+					break;
+				default:
+					if (!Array.isArray(context.itemsByType[i.type])) {
+						context.itemsByType[i.type] = [];
+					}
+					context.itemsByType[i.type].push(i);
 			}
 		}
 
-		skills.sort((a, b) => a.localizedName.localeCompare(b.localizedName));
-
-		context.skills = skills;
+		context.itemsByType.skill.sort(
+			(a, b) => a.localizedName.localeCompare(b.localizedName)
+		);
 
 		let clothing = apparel.filter(a => a.system.appareltype === "clothing");
 		let outfit = apparel.filter(a => a.system.appareltype === "outfit");
@@ -261,18 +245,6 @@ export default class FalloutActorSheet extends ActorSheet {
 			{ apparelType: "armor", list: robotArmor },
 		];
 
-		context.ammo = ammo;
-		context.apparel_mods = apparel_mods;
-		context.books_and_magz = books_and_magz;
-		context.consumables = consumables;
-		context.diseases = diseases;
-		context.miscellany = miscellany;
-		context.perks = perks;
-		context.robot_mods = robot_mods;
-		context.specialAbilities = specialAbilities;
-		context.weapon_mods = weapon_mods;
-		context.weapons = weapons;
-
 		// WRAP INVENTORY DEPENDING ON THE CHARACTER TYPE:
 		// for example put apparel in inventory for all except the character actor.
 
@@ -284,7 +256,7 @@ export default class FalloutActorSheet extends ActorSheet {
 			});
 		}
 		if (this.actor.type === "character") {
-			context.inventory = [...robotApparel, ...robot_mods];
+			context.inventory = [...robotApparel, ...context.itemsByType.robot_mod];
 		}
 		if (this.actor.type === "robot") {
 			context.inventory = [...apparel];